+++ lib/lightning_web/controllers/oidc_html.ex
@@ -1,0 +1,7 @@
+defmodule LightningWeb.OidcHTML do
+  @moduledoc false
+
+  use LightningWeb, :html
+
+  embed_templates "oidc_html/*"
+end